LangSmith-tutorials/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
faiss-cpu==1.12.0
filelock==3.19.1
frozenlist==1.7.0
fsspec==2025.7.0
greenlet==3.2.4
groq==0.31.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.74
langchain-groq==0.3.7
langchain-huggingface==0.3.1
langchain-text-splitters==0.3.9
langgraph==0.6.5
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.2
langsmith==0.4.14
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.6.4
mypy_extensions==1.1.0
networkx==3.5
numpy==2.3.2
orjson==3.11.2
ormsgpack==1.10.0
packaging==25.0
pillow==11.3.0
propcache==0.3.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pypdf==6.0.0
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.5
requests-toolbelt==1.0.0
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
sentence-transformers==5.1.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.4
torch==2.8.0
tqdm==4.67.1
transformers==4.55.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
xxhash==3.5.0
yarl==1.20.1
zstandard==0.24.0
